Rename assents folder to assets
