+++ Front-end/src/pages/Logon/index.js
@@ -6,8 +6,8 @@ import api from '../../services/api'
 
 import './style.css'
 
-import logoImg from '../../assents/logo.svg'
-import heroesImg from '../../assents/heroes.png'
+import logoImg from '../../assets/logo.svg'
+import heroesImg from '../../assets/heroes.png'
 
 export default function Logon() {
     const [id, setId] = useState('')
