+++ Front-end/src/pages/Register/index.js
@@ -7,7 +7,7 @@ import { FiArrowLeft } from 'react-icons/fi'
 import api from '../../services/api'
 import './style.css'
 
-import logoImg from '../../assents/logo.svg'
+import logoImg from '../../assets/logo.svg'
 
 export default function Register() {
     const [name, setName] = useState('')
